Revert to v1.1.37

+++ xsltproc/testThreads.c
@@ -100,7 +100,7 @@ threadRoutine1(void *data)
     xmlChar *result;
     int len;
     xsltStylesheetPtr cur;
-    int id = (int) (size_t) data;
+    int id = (int)(unsigned long) data;
 
     input = xmlReadMemory(doc, strlen(doc), "doc.xml", NULL, 0);
     if (input == NULL) {
@@ -205,7 +205,7 @@ main(void)
 
 	for (i = 0; i < num_threads; i++) {
 	    ret = pthread_create(&tid[i], NULL, threadRoutine1,
-                                 (void *) (size_t) i);
+                                 (void *) (unsigned long) i);
 	    if (ret != 0) {
 		perror("pthread_create");
 		exit(1);
